Moved UI utility to the UiHelper and made documentation for it.
