+++ src/main/java/org/visab/newgui/UiHelper.java
@@ -3,7 +3,13 @@ package org.visab.newgui;
 import org.visab.globalmodel.Vector2;
 
 import javafx.application.Platform;
+import javafx.scene.effect.ColorAdjust;
+import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
+import javafx.scene.image.PixelReader;
+import javafx.scene.image.PixelWriter;
+import javafx.scene.image.WritableImage;
+import javafx.scene.paint.Color;
 
 /**
  * Class containing Helper methods for modifying the Ui.
@@ -18,6 +24,56 @@ public final class UiHelper {
      */
     public static void inovkeOnUiThread(Runnable runnable) {
         Platform.runLater(runnable);
+    }
+    
+    /**
+     * This method recolors any input image pixel by pixel with a given 
+     * JavaFX color.
+     * 
+     * @param inputImage	the image that needs to be recolored.
+     * @param newColor		the new color for the given image.
+     * @return 				the recolored image.
+     */
+    public static Image recolorImage(Image inputImage, Color newColor) {
+        final double r = newColor.getRed();
+        final double g = newColor.getGreen();
+        final double b = newColor.getBlue();
+        final int w = (int) inputImage.getWidth();
+        final int h = (int) inputImage.getHeight();
+        final WritableImage outputImage = new WritableImage(w, h);
+        final PixelWriter writer = outputImage.getPixelWriter();
+        final PixelReader reader = inputImage.getPixelReader();
+        for (int y = 0; y < h; y++) {
+            for (int x = 0; x < w; x++) {
+                // Keeping the opacity of every pixel as it is.
+                writer.setColor(x, y, new Color(r, g, b, reader.getColor(x, y).getOpacity()));
+            }
+        }
+        return outputImage;
+    }
+
+    /**
+     * This method grey scales an image so that other colored elements can be better seen on it.
+     * 
+     * @param inputImage	the image that needs to be grey-scaled.
+     * @return				the grey-scaled image.
+     */
+    public static ImageView greyScaleImage(Image inputImage) {
+        ColorAdjust monochrome = new ColorAdjust();
+        monochrome.setSaturation(-1);
+        ImageView grayScaledImage = new ImageView(inputImage);
+        grayScaledImage.setEffect(monochrome);
+        return grayScaledImage;
+    }
+
+    /**
+     * This method translates a hex color code to a JavaFX color.
+     * 
+     * @param hexCode	the hex color code.
+     * @return			the equivalent JavaFX color.
+     */
+    public static Color translateHexToRgbColor(String hexCode) {
+        return Color.valueOf(hexCode);
     }
 
     /**
